frontend/api/requirements.txt
fastapi
uvicorn
python-multipart
pandas
pdfplumber
pydantic
supabase
python-dotenv
